+++ objectbox-java/src/main/java/io/objectbox/query/PropertyQueryConditionImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2024 ObjectBox Ltd. All rights reserved.
+ * Copyright 2020-2025 ObjectBox Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -375,7 +375,11 @@ public abstract class PropertyQueryConditionImpl<T> extends QueryConditionImpl<T
         private final StringOrder order;
 
         public enum Operation {
-            CONTAINS_KEY_VALUE
+            EQUAL_KEY_VALUE,
+            GREATER_KEY_VALUE,
+            GREATER_EQUALS_KEY_VALUE,
+            LESS_KEY_VALUE,
+            LESS_EQUALS_KEY_VALUE
         }
 
         public StringStringCondition(Property<T> property, Operation op, String leftValue, String rightValue, StringOrder order) {
@@ -388,8 +392,92 @@ public abstract class PropertyQueryConditionImpl<T> extends QueryConditionImpl<T
 
         @Override
         void applyCondition(QueryBuilder<T> builder) {
-            if (op == Operation.CONTAINS_KEY_VALUE) {
-                builder.containsKeyValue(property, leftValue, rightValue, order);
+            if (op == Operation.EQUAL_KEY_VALUE) {
+                builder.equalKeyValue(property, leftValue, rightValue, order);
+            } else if (op == Operation.GREATER_KEY_VALUE) {
+                builder.greaterKeyValue(property, leftValue, rightValue, order);
+            } else if (op == Operation.GREATER_EQUALS_KEY_VALUE) {
+                builder.greaterOrEqualKeyValue(property, leftValue, rightValue, order);
+            } else if (op == Operation.LESS_KEY_VALUE) {
+                builder.lessKeyValue(property, leftValue, rightValue, order);
+            } else if (op == Operation.LESS_EQUALS_KEY_VALUE) {
+                builder.lessOrEqualKeyValue(property, leftValue, rightValue, order);
+            } else {
+                throw new UnsupportedOperationException(op + " is not supported with two String values");
+            }
+        }
+    }
+
+    public static class StringLongCondition<T> extends PropertyQueryConditionImpl<T> {
+        private final Operation op;
+        private final String leftValue;
+        private final long rightValue;
+
+        public enum Operation {
+            EQUAL_KEY_VALUE,
+            GREATER_KEY_VALUE,
+            GREATER_EQUALS_KEY_VALUE,
+            LESS_KEY_VALUE,
+            LESS_EQUALS_KEY_VALUE
+        }
+
+        public StringLongCondition(Property<T> property, Operation op, String leftValue, long rightValue) {
+            super(property);
+            this.op = op;
+            this.leftValue = leftValue;
+            this.rightValue = rightValue;
+        }
+
+        @Override
+        void applyCondition(QueryBuilder<T> builder) {
+            if (op == Operation.EQUAL_KEY_VALUE) {
+                builder.equalKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.GREATER_KEY_VALUE) {
+                builder.greaterKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.GREATER_EQUALS_KEY_VALUE) {
+                builder.greaterOrEqualKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.LESS_KEY_VALUE) {
+                builder.lessKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.LESS_EQUALS_KEY_VALUE) {
+                builder.lessOrEqualKeyValue(property, leftValue, rightValue);
+            } else {
+                throw new UnsupportedOperationException(op + " is not supported with two String values");
+            }
+        }
+    }
+
+    public static class StringDoubleCondition<T> extends PropertyQueryConditionImpl<T> {
+        private final Operation op;
+        private final String leftValue;
+        private final double rightValue;
+
+        public enum Operation {
+            EQUAL_KEY_VALUE,
+            GREATER_KEY_VALUE,
+            GREATER_EQUALS_KEY_VALUE,
+            LESS_KEY_VALUE,
+            LESS_EQUALS_KEY_VALUE
+        }
+
+        public StringDoubleCondition(Property<T> property, Operation op, String leftValue, double rightValue) {
+            super(property);
+            this.op = op;
+            this.leftValue = leftValue;
+            this.rightValue = rightValue;
+        }
+
+        @Override
+        void applyCondition(QueryBuilder<T> builder) {
+            if (op == Operation.EQUAL_KEY_VALUE) {
+                builder.equalKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.GREATER_KEY_VALUE) {
+                builder.greaterKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.GREATER_EQUALS_KEY_VALUE) {
+                builder.greaterOrEqualKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.LESS_KEY_VALUE) {
+                builder.lessKeyValue(property, leftValue, rightValue);
+            } else if (op == Operation.LESS_EQUALS_KEY_VALUE) {
+                builder.lessOrEqualKeyValue(property, leftValue, rightValue);
             } else {
                 throw new UnsupportedOperationException(op + " is not supported with two String values");
             }